check: extract typings  of check* methods
