--- check/check-all.js
@@ -1,8 +1,0 @@
-const check = require('./check');
-module.exports = (fields, message) => check(fields, [
-  'body',
-  'query',
-  'cookies',
-  'headers',
-  'params'
-], message);
