--- check/check-params.js
@@ -1,2 +1,0 @@
-const check = require('./check');
-module.exports = (fields, message) => check(fields, ['params'], message);
