--- check/check-query.js
@@ -1,2 +1,0 @@
-const check = require('./check');
-module.exports = (fields, message) => check(fields, ['query'], message);
