+++ check/validation-chain-builders.spec.ts
@@ -1,0 +1,9 @@
+import { check, body, cookie, header, param, query } from './validation-chain-builders';
+import { ValidationChain } from './validation-chain';
+
+let chain: ValidationChain = check('foo');
+chain = body('foo');
+chain = cookie('foo');
+chain = header('foo');
+chain = param('foo');
+chain = query('foo');
