+++ check/validation-chain.d.ts
@@ -1,0 +1,10 @@
+import * as express from 'express';
+import { Validator } from '../shared-typings';
+
+export interface ValidationChain extends express.RequestHandler, Validator {
+  custom(validator: CustomValidator): this;
+}
+
+export interface CustomValidator {
+  (value: any, options: { req: express.Request, location: string, path: string }): any;
+}
